+++ .gitignore
@@ -1,9 +1,8 @@
 dist
 src/components/content/conditions-table-loader.tsx
 src/styles/tailwind-gen.css
+tailwind-gen.theme.cjs
 scripts/prefixer/tmp/
 node_modules
 .env
 .DS_Store
-tailwind.theme.js
-tailwind.theme.cjs
