+++ package.json
@@ -20,8 +20,8 @@
     "dev": "run-p dev:*",
     "dev:vite": "vite",
     "dev:css": "npm run generate:css -- --watch",
-    "generate:css": "npm run generate-theme-cjs && tailwindcss -c ./tailwind.config.cjs -i ./src/styles/tailwind.css -o ./src/styles/tailwind-gen.css",
-    "generate-theme-cjs": "tsc tailwind.theme.ts --target es5 --module commonjs && mv tailwind.theme.js tailwind.theme.cjs",
+    "generate:css": "npm run generate-theme-cjs && tailwindcss -i ./src/styles/tailwind.css -o ./src/styles/tailwind-gen.css",
+    "generate-theme-cjs": "tsc ./src/styles/tailwind.theme.ts --target es5 --module commonjs && mv ./src/styles/tailwind.theme.js ./tailwind-gen.theme.cjs",
     "build": "npm run generate:css && run-p build:*",
     "build:scripts": "vite build",
     "build:types": "tsc --emitDeclarationOnly && tsc-alias",
