+++ tsconfig.json
@@ -21,11 +21,9 @@
     "paths": {
       "@/*": ["src/*"]
     },
-    "typeRoots": ["./typings"],
   },
   "include": [
     "src",
     "scripts"
   ],
-  "exclude": ["./typings"],
 }
